Create analytics package in models package and move relevant classes
