+++ src/main/java/no/oslomet/aaas/model/analytics/AnonymizationResultPayload.java
@@ -1,6 +1,4 @@
-package no.oslomet.aaas.model;
-
-import java.util.Map;
+package no.oslomet.aaas.model.analytics;
 
 
 /***
