+++ src/test/java/no/oslomet/aaas/GenerateTestData.java
@@ -1,6 +1,9 @@
 package no.oslomet.aaas;
 
 import no.oslomet.aaas.model.*;
+import no.oslomet.aaas.model.analytics.DistributionOfRisk;
+import no.oslomet.aaas.model.analytics.ReIdentificationRisk;
+import no.oslomet.aaas.model.analytics.RiskProfile;
 import no.oslomet.aaas.utils.ARXDataFactory;
 import no.oslomet.aaas.utils.DataFactory;
 import org.deidentifier.arx.Data;
