+++ src/test/java/no/oslomet/aaas/analyzer/ARXReIdentificationRiskFactoryTest.java
@@ -1,7 +1,6 @@
 package no.oslomet.aaas.analyzer;
 
-import no.oslomet.aaas.analyzer.ARXReIdentificationRiskFactory;
-import no.oslomet.aaas.model.ReIdentificationRisk;
+import no.oslomet.aaas.model.analytics.ReIdentificationRisk;
 import org.deidentifier.arx.*;
 import org.deidentifier.arx.criteria.KAnonymity;
 import org.junit.Before;
