+++ src/test/java/no/oslomet/aaas/service/AnonymizationServiceTest.java
@@ -4,6 +4,7 @@ import no.oslomet.aaas.GenerateTestData;
 import no.oslomet.aaas.analyzer.ARXAnalyzer;
 import no.oslomet.aaas.anonymizer.ARXAnonymizer;
 import no.oslomet.aaas.model.*;
+import no.oslomet.aaas.model.analytics.AnonymizationResultPayload;
 import no.oslomet.aaas.utils.*;
 import org.junit.Assert;
 import org.junit.Before;
